add STI to LinkedAccount   - also fixes some minor issues (rubocop)
