+++ app/resources/profile_link_resource.rb
@@ -1,0 +1,7 @@
+class ProfileLinkResource < BaseResource
+  immutable
+  attribute :url
+
+  has_one :user
+  has_one :profile_link_site
+end
