+++ app/resources/user_resource.rb
@@ -19,6 +19,7 @@ class UserResource < BaseResource
   has_many :following
   has_many :blocks
   has_many :linked_accounts
+  has_many :profile_links
   has_many :media_follows
   has_many :user_roles
   has_many :library_entries
