+++ config/routes.rb
@@ -605,6 +605,7 @@ Rails.application.routes.draw do
     jsonapi_resources :roles
     jsonapi_resources :user_roles
     jsonapi_resources :reports
+    jsonapi_resources :linked_accounts
     resources :activities, only: %i[destroy]
     get '/feeds/:group/:id', to: 'feeds#show'
     post '/feeds/:group/:id/_read', to: 'feeds#mark_read'
